Fix for Issue 1770 on 3.1 (#1771)
* [CI] adding action for clang compiler
* [C] Ingnoring postgres warnings
* [dijkstra] compiling with clang
* [common] compiling with clang
* [astar] compiles with clang
* [ksp] compiles with clang
* [tsp] compiles with clang
* [alpha_shape] compiles with clang
* [tspEuclidean] compiles with clang
* [witPoints] compiles with clang
* [bellman_ford] compiles with clang
* [chinese] compiles with clang
* [transitiveClosure] compiles with clang
* [breadthFirstSearch] compiles with clang
* [components] compiles with clang
* [maxflow] pgr_maxflow.hpp compiles with clang
* [maxflow] compiles with clang
* [lineGraph] compiles with clang
* [bdAstar] compiles with clang
* [bdDijkstra] compiles with clang
* [trsp] compiles with clang
* [pickDeliver] compiles with clang
* [dijkstra][docqueries] giving a new correct alternate result
* Removing a test of a function that does not exist on pgRouting
* removing internal query tests that were created when developing contraction and are not used
* [tsp][docqueries] Modification of rand() gives new results
* roll back changes that affect ksp
* [C++] moving pgr_messages.cpp to cpp_common
* roll back changes that affect trsp
* src/common/basePath_SSEC.cpp compiles with clang
* src/trsp/pgr_trspHandler.cpp compiles with clang
* [ci][pgtap] adding testing after clang compilation
* [doc] adding NEWS & Release notes
Also updating compilation_time and compiler & hash on pgr_full_version
* [build] rolling back build-extension-update-files1.pl
* [C++] adding some minimal changes from the comments of the PR

+++ src/trsp/GraphDefinition.cpp
@@ -79,7 +79,7 @@ void GraphDefinition::deleteall() {
 double GraphDefinition::construct_path(int64 ed_id, int64 v_pos) {
     if (parent[ed_id].ed_ind[v_pos] == -1) {
         path_element_tt pelement;
-        GraphEdgeInfo* cur_edge = m_vecEdgeVector[ed_id];
+        GraphEdgeInfo* cur_edge = m_vecEdgeVector[static_cast<size_t>(ed_id)];
         if (v_pos == 0) {
             pelement.vertex_id = cur_edge->m_lStartNode;
             pelement.cost = cur_edge->m_dCost;
@@ -95,7 +95,7 @@ double GraphDefinition::construct_path(int64 ed_id, int64 v_pos) {
     double ret = construct_path(parent[ed_id].ed_ind[v_pos],
         parent[ed_id].v_pos[v_pos]);
     path_element_tt pelement;
-    GraphEdgeInfo* cur_edge = m_vecEdgeVector[ed_id];
+    GraphEdgeInfo* cur_edge = m_vecEdgeVector[static_cast<size_t>(ed_id)];
     if (v_pos == 0) {
         pelement.vertex_id = cur_edge->m_lStartNode;
         pelement.cost = m_dCost[ed_id].endCost - ret;  // cur_edge.m_dCost;
@@ -134,11 +134,11 @@ double GraphDefinition::getRestrictionCost(
                 flag = false;
                 break;
             }
-            if (precedence != m_vecEdgeVector[edge_ind]->m_lEdgeID) {
+            if (precedence != m_vecEdgeVector[static_cast<size_t>(edge_ind)]->m_lEdgeID) {
                 flag = false;
                 break;
             }
-            auto parent_ind = parent[edge_ind].ed_ind[v_pos];
+            auto parent_ind = parent[static_cast<size_t>(edge_ind)].ed_ind[static_cast<size_t>(v_pos)];
             v_pos = parent[edge_ind].v_pos[v_pos];
             edge_ind = parent_ind;
         }
@@ -159,7 +159,7 @@ void GraphDefinition::explore(
     std::greater<PDP> > &que) {
     double totalCost;
     for (const auto &index : vecIndex) {
-        auto new_edge = m_vecEdgeVector[index];
+        auto new_edge = m_vecEdgeVector[static_cast<size_t>(index)];
         double extCost = 0.0;
         if (m_bIsturnRestrictOn) {
             extCost = getRestrictionCost(cur_edge.m_lEdgeIndex,
@@ -223,7 +223,7 @@ int GraphDefinition::my_dijkstra1(
             m_bIsGraphConstructed = true;
         }
         GraphEdgeInfo* start_edge_info =
-        m_vecEdgeVector[m_mapEdgeId2Index[start_edge_id]];
+        m_vecEdgeVector[static_cast<size_t>(m_mapEdgeId2Index[static_cast<int64>(start_edge_id)])];
         edge_t start_edge;
         int64 start_vertex, end_vertex;
         m_dStartpart = start_part;
@@ -261,7 +261,7 @@ int GraphDefinition::my_dijkstra1(
         }
 
     GraphEdgeInfo* end_edge_info =
-    m_vecEdgeVector[m_mapEdgeId2Index[end_edge_id]];
+    m_vecEdgeVector[static_cast<size_t>(m_mapEdgeId2Index[static_cast<int64>(end_edge_id)])];
     edge_t end_edge;
 
     if (end_part == 0.0) {
@@ -339,7 +339,7 @@ int GraphDefinition:: my_dijkstra2(
                 for (const auto &source : vecsource) {
                     temp_precedencelist.clear();
                     temp_precedencelist.push_back(
-                        m_vecEdgeVector[source]->m_lEdgeID);
+                        m_vecEdgeVector[static_cast<size_t>(source)]->m_lEdgeID);
                     m_ruleTable[dest_edge_id].push_back(Rule(rule.first,
                         temp_precedencelist));
                 }
@@ -352,7 +352,7 @@ int GraphDefinition:: my_dijkstra2(
             vecsource = m_mapNodeId2Edge[end_vertex];
             for (const auto &source : vecsource) {
                 m_ruleTable.insert(std::make_pair(
-                    m_vecEdgeVector[source]->m_lEdgeID, tmpRules));
+                    m_vecEdgeVector[static_cast<size_t>(source)]->m_lEdgeID, tmpRules));
             }
         }
     }
@@ -404,7 +404,7 @@ int GraphDefinition:: my_dijkstra3(
     GraphEdgeInfo* cur_edge = NULL;
 
     for (const auto &source : vecsource) {
-        cur_edge = m_vecEdgeVector[source];
+        cur_edge = m_vecEdgeVector[static_cast<size_t>(source)];
         if (cur_edge->m_lStartNode == start_vertex) {
             if (cur_edge->m_dCost >= 0.0) {
                 m_dCost[cur_edge->m_lEdgeIndex].endCost = cur_edge->m_dCost;
@@ -430,7 +430,7 @@ int GraphDefinition:: my_dijkstra3(
         PDP cur_pos = que.top();
         que.pop();
         int64 cured_index = cur_pos.second.first;
-        cur_edge = m_vecEdgeVector[cured_index];
+        cur_edge = m_vecEdgeVector[static_cast<size_t>(cured_index)];
 
         if (cur_pos.second.second) {  // explore edges connected to end node
             cur_node = cur_edge->m_lEndNode;
@@ -482,7 +482,7 @@ int GraphDefinition:: my_dijkstra3(
 
         *path = reinterpret_cast<path_element_tt *>(
                 malloc(sizeof(path_element_tt) * (m_vecPath.size() + 1)));
-        *path_count = static_cast<int>(m_vecPath.size());
+        *path_count = m_vecPath.size();
 
         for (size_t i = 0; i < *path_count; i++) {
             (*path)[i].vertex_id = m_vecPath[i].vertex_id;
@@ -506,7 +506,7 @@ int GraphDefinition:: my_dijkstra3(
 bool GraphDefinition::get_single_cost(double total_cost, path_element_tt **path,
      size_t *path_count) {
     GraphEdgeInfo* start_edge_info =
-    m_vecEdgeVector[m_mapEdgeId2Index[m_lStartEdgeId]];
+    m_vecEdgeVector[static_cast<size_t>(m_mapEdgeId2Index[m_lStartEdgeId])];
     if (m_dEndPart >= m_dStartpart) {
         if (start_edge_info->m_dCost >= 0.0 && start_edge_info->m_dCost *
             (m_dEndPart - m_dStartpart) <= total_cost) {
@@ -603,7 +603,7 @@ bool GraphDefinition::addEdge(edge_t edgeIn) {
     newEdge->m_vecEndConnedtedEdge.clear();
     newEdge->m_vecRestrictedEdge.clear();
     newEdge->m_lEdgeID = edgeIn.id;
-    newEdge->m_lEdgeIndex = m_vecEdgeVector.size();
+    newEdge->m_lEdgeIndex = static_cast<int64>(m_vecEdgeVector.size());
     newEdge->m_lStartNode = edgeIn.source;
     newEdge->m_lEndNode = edgeIn.target;
     newEdge->m_dCost = edgeIn.cost;
@@ -626,11 +626,11 @@ bool GraphDefinition::addEdge(edge_t edgeIn) {
     if (itNodeMap != m_mapNodeId2Edge.end()) {
         // Connect current edge with existing edge with start node
         // connectEdge(
-        int64 lEdgeCount = itNodeMap->second.size();
+        int64 lEdgeCount = static_cast<int64>(itNodeMap->second.size());
         int64 lEdgeIndex;
         for (lEdgeIndex = 0; lEdgeIndex < lEdgeCount; lEdgeIndex++) {
-            int64 lEdge = itNodeMap->second.at(lEdgeIndex);
-            connectEdge(*newEdge, *m_vecEdgeVector[lEdge], true);
+            int64 lEdge = itNodeMap->second.at(static_cast<size_t>(lEdgeIndex));
+            connectEdge(*newEdge, *m_vecEdgeVector[static_cast<size_t>(lEdge)], true);
         }
     }
 
@@ -640,11 +640,11 @@ bool GraphDefinition::addEdge(edge_t edgeIn) {
     if (itNodeMap != m_mapNodeId2Edge.end()) {
         // Connect current edge with existing edge with end node
         // connectEdge(
-        int64 lEdgeCount = itNodeMap->second.size();
+        int64 lEdgeCount = static_cast<int64>(itNodeMap->second.size());
         int64 lEdgeIndex;
         for (lEdgeIndex = 0; lEdgeIndex < lEdgeCount; lEdgeIndex++) {
-            int64 lEdge = itNodeMap->second.at(lEdgeIndex);
-            connectEdge(*newEdge, *m_vecEdgeVector[lEdge], false);
+            int64 lEdge = itNodeMap->second.at(static_cast<size_t>(lEdgeIndex));
+            connectEdge(*newEdge, *m_vecEdgeVector[static_cast<size_t>(lEdge)], false);
         }
     }
